Refactor into own custom elements * input-dropdown * command-palett
