+++ client/ListMapper/list-mapper/src/app.js
@@ -1,7 +1,6 @@
 import { PLATFORM } from 'aurelia-pal';
 import {observable} from 'aurelia-framework';
 
-import hotkeys from 'hotkeys-js';
 import {commandList} from './resources/commandStorage/command-storage'
 import './app.less';
 import 'font-awesome/css/font-awesome.css';
@@ -12,30 +11,12 @@ export class App {
   constructor() {
     this.showNavbar = false;
     this.showCommandPalett = false;
-    
     this.suggestedList = null;
     this.key = 'name'
-    
-    this.commands = [
-      {name: 'Say Hello', id: 'sayHello'},
-      {name: 'Say Bye', id: 'sayBye'},
-    ]
   }
 
   attached() {
-    this.initShortCuts();
     this.simpleCommand = commandList.simpleCommand;
-  }
-
-  initShortCuts() {
-    // Init hotkeys
-    let hotkey = hotkeys.noConflict();
-    hotkeys.filter = function () { return true }; // 2018-08-09 23:30:46 what does this do?
-
-    hotkey('ctrl+;', () => {
-      this.showCommandPalett = !this.showCommandPalett;
-      console.log('yay')
-    });
   }
 
   submitCommand() {
