+++ client/ListMapper/list-mapper/src/resources/common/elements/command-palett.less
@@ -1,0 +1,63 @@
+@import (reference) '../../common/styles/colors.less';
+
+.command-palett-container {
+  background: @pri-clr;
+  
+  // following block centers an absolute element
+  position: absolute;
+    left: 0;
+    right: 0;
+    margin: 0px auto;
+    width: 80vw;
+  
+  z-index: 1;
+
+  .command-palett {
+    background: @pri-clr;
+    border: 1px solid white;
+    justify-content: center;
+    padding: 0 1.5em;
+    z-index: 1;
+
+    input::-webkit-calendar-picker-indicator {
+      display: none;
+    }
+
+    input {
+      background: @pri-clr;
+      color: @grey-clr;
+
+      outline-style: none;
+      border: 0;
+      width: 90%;
+    }
+
+    button {
+      display: none;
+    }
+
+  }
+
+  div {
+
+    div {
+
+      &.item-active {
+        .active();
+      }
+
+      &:hover {
+        .active();
+      }
+
+      // &:first-child {
+      //   .active();
+      // }
+    }
+      
+  }
+}
+
+.active() {
+  background: @pri-clr-truthy-hover;
+}
