+++ client/ListMapper/list-mapper/src/resources/common/elements/input-dropdown.js
@@ -1,0 +1,13 @@
+import {bindable} from 'aurelia-framework';
+
+export class InputDropdown {
+  @bindable value;
+  @bindable simpleCommand;
+  @bindable showCommandPalett;
+  @bindable key;
+
+  submitCommand() {
+    console.log('submit command');
+  }
+}
+
